refactor: use whole width for utxo list in confirm modal
