+++ src/components/Send/SendForm.tsx
@@ -214,7 +214,6 @@ export interface SendFormValues {
   txFee?: TxFee
   isCoinJoin: boolean
   numCollaborators?: number
-  consideredUtxos?: string[]
 }
 
 interface InnerSendFormProps {
